backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
ollama==0.6.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
werkzeug==2.3.7
pytz==2023.3
requests==2.31.0
fastapi==0.129.0
uvicorn==0.40.0
python-multipart==0.0.22
